backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.135.1
uvicorn==0.42.0
starlette==0.52.1

# Database
sqlalchemy==2.0.48
psycopg2-binary==2.9.11

# Data Validation
pydantic==2.12.5
email-validator>=2.0.0

# Authentication & Security
python-jose==3.5.0
PyJWT==2.12.1
google-auth==2.41.1
passlib==1.7.4
bcrypt==5.0.0

# Utilities
python-multipart==0.0.22
requests==2.32.5
typing-extensions==4.15.0
annotated-types==0.7.0

# Dependencies
greenlet==3.3.2
click==8.3.1
h11==0.16.0
ecdsa==0.19.1
rsa==4.9.1
pyasn1==0.6.3
six==1.17.0
idna==3.11
anyio==4.12.1
colorama==0.4.6
pydantic-core==2.41.5
typing-inspection==0.4.2
annotated-doc==0.0.4
